requirements.txt
appnope==0.1.3
    # via ipython
asgiref==3.5.2
    # via django
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via jsonschema
backcall==0.2.0
    # via ipython
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via pip-tools
decorator==5.1.1
    # via ipython
django==4.0.4
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
    #   drf-spectacular
django-cors-headers==3.12.0
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-ipware==4.0.2
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.22.1
    # via -r requirements.in
executing==0.8.3
    # via stack-data
funcy==1.17
    # via -r requirements.in
humanize==4.1.0
    # via -r requirements.in
idna==3.3
    # via requests
inflection==0.5.1
    # via drf-spectacular
ipython==8.3.0
    # via -r requirements.in
jedi==0.18.1
    # via ipython
jsonschema==4.5.1
    # via drf-spectacular
matplotlib-inline==0.1.3
    # via ipython
parso==0.8.3
    # via jedi
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.6.1
    # via -r requirements.in
prompt-toolkit==3.0.29
    # via ipython
psycopg2-binary==2.9.3
    # via -r requirements.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.12.0
    # via ipython
pyrsistent==0.18.1
    # via jsonschema
pytz==2022.1
    # via djangorestframework
pyyaml==6.0
    # via drf-spectacular
requests==2.27.1
    # via -r requirements.in
six==1.16.0
    # via asttokens
sqlparse==0.4.2
    # via django
stack-data==0.2.0
    # via ipython
tomli==2.0.1
    # via pep517
traitlets==5.2.1.post0
    # via
    #   ipython
    #   matplotlib-inline
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.9
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
whatever==0.6
    # via -r requirements.in
wheel==0.37.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
